Modifications improvement

+++ src/main/java/com/ecommerence/platform/service/ProductService.java
@@ -1,8 +1,10 @@
 package com.ecommerence.platform.service;
 
+import com.ecommerence.platform.constants.AppConstants;
 import com.ecommerence.platform.entity.Product;
 import com.ecommerence.platform.enums.DirectionEnum;
 import com.ecommerence.platform.enums.ProductOrderEnum;
+import com.ecommerence.platform.exception.ProductNotFoundException;
 import com.ecommerence.platform.repository.ProductRepository;
 import com.ecommerence.platform.response.ProductsResponse;
 import com.ecommerence.platform.util.SortRequestBuilder;
@@ -16,31 +18,41 @@ import java.util.Date;
 @Service
 public class ProductService {
 
-    ProductRepository productRepository;
+    private final ProductRepository productRepository;
 
     public ProductService(ProductRepository productRepository) {
         this.productRepository = productRepository;
     }
 
-    public String createProduct(Product product) {
+    public Product createProduct(Product product) {
 
         product.setCreatedDate(new Date());
+
         productRepository.save(product);
 
-        return "Success";
+        return product;
     }
 
-    public String deleteProduct(Integer id) {
+    public void deleteProduct(Integer id) throws ProductNotFoundException {
+
+        if(!productRepository.existsById(id) ){
+            throw new ProductNotFoundException(String.format(AppConstants.PRODUCT_WITH_ID_NOT_FOUND_MESSAGE_TEMPLATE, id));
+        }
         productRepository.deleteById(id);
-
-        return "Success";
     }
 
-    public String updateProduct(Product product) {
+    public Product updateProduct(Integer id, Product product) throws ProductNotFoundException {
+
+        if(!productRepository.existsById(id) ){
+           throw new ProductNotFoundException(String.format(AppConstants.PRODUCT_WITH_ID_NOT_FOUND_MESSAGE_TEMPLATE, id));
+        }
+
+        product.setId(id);
         product.setUpdatedDate(new Date());
+
         productRepository.save(product);
 
-        return "Success";
+        return product;
     }
 
     public ProductsResponse getProducts(ProductOrderEnum orderBy, DirectionEnum direction, Integer page, Integer pageSize) {
